+++ client/src/BookProfile/services/BookProfileService.js
@@ -1,26 +1,11 @@
+import { commonService } from '../../Common/services/CommonService';
 const API_URL = process.env.REACT_APP_API_URL;
-
-const checkStatus = (response) => {
-    if(!response.ok) {
-        throw Error(response.status);
-    }
-    return response;
-}
-
-const getToken = () => {
-    let token = window.localStorage.getItem('user');
-    token = JSON.parse(token);
-    token = token.token;
-
-    return token;
-}
-
 
 const getBookById = (id) => {
     const url = `${API_URL}/books/${id}`;
 
     let myHeaders = new Headers();
-    myHeaders.append("Authorization", `Bearer ${getToken()}`);
+    myHeaders.append("Authorization", `Bearer ${commonService.getToken()}`);
 
     let fetchData = {
         method: 'GET',
@@ -30,7 +15,7 @@ const getBookById = (id) => {
     };
 
     return fetch(url, fetchData)
-        .then(response => checkStatus(response))
+        .then(response => commonService.checkStatus(response))
         .then(response => response.text())
         .then(response => JSON.parse(response))
         .catch(error => {
